@@ -1,6 +1,6 @@
-var express = require('express');
-var path = require('path');
-var open = require('opn');
+import express from 'express';
+import path from 'path';
+import open from 'opn';
 
 var port = 3000;
 var app = express();
@@ -9,10 +9,12 @@ app.get('/', function(req, res) {
   res.sendFile(path.join(__dirname,'../src/index.html'));
 });
 
-app.listen(port, function(err) {
-  if (err){
+app.listen(port, function(err){
+  if(err){
     console.log(err);
   } else {
     open('http://localhost:' + port);
   }
 });
+
+
